requirements.txt
streamlit==1.36.0
PyPDF2==3.0.1
sentence-transformers==2.6.1
numpy==1.26.4
genai==2.1.0
google-generativeai==0.5.4
